// src/server/models/inventories/dto/item/delete.item.schema.ts
/**
 * administrare - web platform for internal data management
 * Copyright (C) 2022 imperatoria
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */

import { MutasiBarangKeluar, MutasiBarangMasuk, Saldo, SaldoAkhir } from "../../schema/inventory.schema";

export class ParameterDeleteItemDto {
    tahun: number;
    kategori_id: number;
    barang_id: number;
}

export class ResponseDeleteItemDto {
    id: number;
    nama: string;
    satuan: string;
    saldo: Saldo;
    mutasi_barang_masuk: MutasiBarangMasuk;
    mutasi_barang_keluar: MutasiBarangKeluar;
    saldo_akhir: SaldoAkhir;
}